Adding imgui code and fixing another ellapsed time code....

+++ src/chip8.cpp
@@ -92,7 +92,7 @@ void Chip8::runStep() {
     }
 
     // sleep according to the clock
-    float time_to_sleep = (1000.0/clock)*1000 - (int)ellapsed_fetch.count()*1000000;
+    float time_to_sleep = (1000.0/clock)*1000 - ellapsed_fetch.count()*1000000;
     if (time_to_sleep > 0) {
         std::this_thread::sleep_for(std::chrono::microseconds(int(time_to_sleep)));
     }
